requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
Pillow==10.1.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-dotenv==1.0.0
django-allauth==0.55.0 